Updated get_acc_list.py.
Updates include using argparse for command line arguments and adding
new flags. Arguments include number of underscore-separated "units"
within the infile name to include in the output name, the number of
header lines to skip at the start of the file, and the column number
that the relevant accessions are placed on.